remove test link_coin
