--- tests/integrate_test/test_integrate_builtin_governance.py
@@ -1,47 +1,0 @@
-# -*- coding: utf-8 -*-
-
-# Copyright 2018 ICON Foundation
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-# http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-"""IconScoreEngine testcase
-"""
-
-import unittest
-
-from iconservice.base.address import ZERO_SCORE_ADDRESS
-from iconservice.icon_constant import ConfigKey
-from tests.integrate_test.test_integrate_base import TestIntegrateBase
-
-
-class TestIntegrateBuiltinGovernance(TestIntegrateBase):
-
-    def _make_init_config(self) -> dict:
-        return {ConfigKey.SERVICE: {ConfigKey.SERVICE_SCORE_PACKAGE_VALIDATOR: True}}
-
-    def test_builtin_governance(self):
-
-        tx1 = self._make_deploy_tx("test_scores",
-                                   "l_coin_0_5_0",
-                                   self._addr_array[0],
-                                   ZERO_SCORE_ADDRESS)
-
-        prev_block, tx_results = self._make_and_req_block([tx1])
-
-        self._write_precommit_state(prev_block)
-
-        self.assertEqual(tx_results[0].status, int(True))
-
-
-if __name__ == '__main__':
-    unittest.main()
