--- tests/integrate_test/test_samples/test_scores/l_coin_0_5_0/package.json
@@ -1,5 +1,0 @@
-{
-    "version": "0.0.1",
-    "main_file": "link_coin",
-    "main_score": "LinkCoin"
-}
