--- tests/integrate_test/test_samples/test_scores/l_coin_0_5_0_using_import_os/__init__.py
@@ -1,1 +1,0 @@
-from .link_coin import LinkCoin # noqa
