--- tests/integrate_test/test_samples/test_scores/l_coin_0_5_0_using_import_os/ownership/ownable.py
@@ -1,17 +1,0 @@
-from iconservice import *
-
-
-def only_owner(func):
-    if not isfunction(func):
-        revert(f"{func} isn't function.")
-
-    @wraps(func)
-    def __wrapper(calling_obj: object, *args, **kwargs):
-        if isinstance(calling_obj, IconScoreBase):
-            sender = calling_obj.tx.origin if calling_obj.tx else calling_obj.msg.sender
-            if calling_obj.owner != sender:
-                revert(f"{sender} don't have authority.")
-
-            return func(calling_obj, *args, **kwargs)
-
-    return __wrapper
