+++ tests/integrate_test/test_samples/test_scores/test_score_using_import_os/package.json
@@ -1,0 +1,5 @@
+{
+    "version": "0.0.1",
+    "main_file": "test_score_using_import_os",
+    "main_score": "TestScoreUsingImportOs"
+}
