+++ tests/integrate_test/test_samples/test_scores/test_score_using_import_os/test_score_using_import_os.py
@@ -1,0 +1,18 @@
+from iconservice import *
+import os
+
+class TestScoreUsingImportOs(IconScoreBase):
+
+    def __init__(self, db: IconScoreDatabase) -> None:
+        super().__init__(db)
+
+    def on_install(self, value: 'Address' = None, value1: Address = None, value2: int = None) -> None:
+        super().on_install()
+
+    def on_update(self) -> None:
+        super().on_update()
+
+    @external(readonly=True)
+    def hello(self) -> int:
+        return 0
+
